Factor out s3 client method
