+++ pull_scan_metadata/src/python/scan_metadata_app/run.py
@@ -9,6 +9,7 @@ from flywheel_gear_toolkit import GearToolkitContext
 from inputs.api_key import get_api_key
 from inputs.context_parser import parse_config
 from inputs.parameter_store import get_parameter_store
+from s3.s3_client import get_s3_client
 from scan_metadata_app.main import run
 from ssm_parameter_store import EC2ParameterStore
 
@@ -58,26 +59,4 @@ if __name__ == "__main__":
     main()
 
 
-def get_s3_client(*, store: EC2ParameterStore, path: str):
-    """Returns the S3 client for the access credentials in the parameter store
-    at the given path.
 
-    Args:
-      store: the parameter store with the credentials
-      path: the parameter store path for the S3 credentials
-    Returns:
-      the boto3 S3 client
-    """
-    # Get S3 credentials
-    parameters = store.get_parameters_by_path(path=path)
-    access_key = parameters.get('accesskey')
-    secret_key = parameters.get('secretkey')
-    region = parameters.get('region')
-
-    # Initialize the S3 client
-    client = boto3.client('s3',
-                          aws_access_key_id=access_key,
-                          aws_secret_access_key=secret_key,
-                          region_name=region)
-
-    return client
